FacialRecognition/FacialRecognition/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
dlib==19.24.2
